% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsinf.R
\name{lsinf}
\alias{lsinf}
\title{Returns elements of the scaled fisher information matrix for location-scale distributions}
\usage{
lsinf(z, censor.type, distribution)
}
\arguments{
\item{z}{standardized time}

\item{censor.type}{A characcter string: uncensored, left, or right}

\item{distribution}{A characcter string: sev, normal, lev, logistic.}
}
\value{
A list containing vectors f11, f12 and f22 of Fisher information matrix elements.
}
\description{
Returns elements of the scaled fisher information matrix for location-scale distributions
}
\examples{
lsinf(0.5, "right", "sev")
lsinf(0.5, "left", "normal"
lsinf(seq(-1, 1, by=.1), "right", "sev")
lsinf(seq(-2, 2, by=.2), "right", "normal")
}
